+++ include/kernel/string.h
@@ -26,6 +26,7 @@
 /* Memory functions */
 void* memset(void* dest, int val, size_t len);
 void* memcpy(void* dest, const void* src, size_t len);
+void* memmove(void* dest, const void* src, size_t len);
 int memcmp(const void* s1, const void* s2, size_t n);
 
 /* String functions */